lib/python/briar/requirements.txt
dlib>=19.18.0
grpcio>=1.34.0
grpcio-tools>=1.34.0
protobuf
pandas
mypy-protobuf
opencv-python>=4.1.2.30
numpy
tqdm
pandas
matplotlib
colorama
imutils
pillow
imageio
imageio-ffmpeg
scipy
scikit-learn
scikit-image
pyvision-toolkit


